Remove temporary refactored files
